requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-auth==2.10.0
google-pasta==0.2.0
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.4
httplib2==0.20.2
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scikit-learn-extra==0.2.0
scipy==1.9.0
six==1.16.0
sklearn==0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.0
transformers==4.21.1
typing-extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
Pillow==9.0.1
reportlab==3.6.8
gTTS==2.3.0
pdf2image==1.16.0
newspaper3k==0.2.8
docx2txt==0.8
azure-cognitiveservices-speech==1.25.0
python-dotenv==0.21.1
google-api-python-client==2.80.0
